chore: refactor integration tests
Asserting over the expectedSqlTerms and dependencies while useful for porting things from pgdelta
and debugging. Considerably slow down developping when changing things around.

Asserting that the final diffscript is a valid sql, then that the result over the target database schema
catalogs are the same is enough for most of the tests functional validity and makes for much more resilient
and less maintain burden.

Now we should only add a few tests that assert over the actual produced sql for styling purpose or specific assertion